requirements.txt
alembic==0.9.6
backports-abc==0.5
banal==0.2.0
certifi==2017.7.27.1
chardet==3.0.4
dataset==1.0.2
functools32==3.2.3.post2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
normality==0.5.0
pyserial==3.4
python-dateutil==2.6.1
python-editor==1.0.3
python-gsmmodem-new==0.12
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
tornado==4.5.2
Tornado-JSON==1.3.2
